code to trigger and test what happens in race condition
Not at a real endpoint, but I can trigger a race condition where
multiple set statements for the same document collide.

Interesting to see where the error is handled.  I can actually handle
and continue to process lots of errors if the catch handler inside the
loop is uncommented.  Need to write this up properly I think.