mid hack, but I've figured out the problem
The problem is that couchdb is asynchronous.  It returns values at a
rate that is completely independent of my node.js code, so I can't
expect the retries to nicely count down from 10 to 0.  In practice,
some get/put sequences sneak in after another doc has successfully get
and put, so there is no conflict and no decrement.

repeatedly running the test will generate erratic totals of successful
puts and failed puts.