fix disconnect betwn local and travis tests
local test was using two docker containers, and couchdb is not on
localhost, so the default value for "host" of '127.0.0.1' was
resulting in connection refused error, not the correct Not Found
error.

To fix, have to pass the 'host' param from the config file.

For some reason, it looks like I don't need to use auth anywhere in
this library.  I added the necessary bits, but it didn't seem to make
a diff.

Anyway, fixed it.